Add functionality to decide on which room we are in

+++ app/src/main/java/com/example/example6/Rectangle.java
@@ -59,7 +59,7 @@ public class Rectangle {
         return length;
     }
 
-    public double getRoom() {
+    public int getRoom() {
         return room;
     }
 }
